+++ Debug/AppDelegateMain.swift
@@ -1,0 +1,5 @@
+import UIKit
+
+@UIApplicationMain
+final class AppDelegateMain: AppDelegate {
+}