+++ CHANGELOG.md
@@ -4,13 +4,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2019-09-05
+### Added
+- iOS 13.0 support.
+- [Bazel](https://bazel.build) support.
+- [Tuist](https://tuist.io) project for easier debugging.
+
+### Removed
+- `fatalError`
+- `AppDelegate` class initializers.
+
 ## [0.1.1] - 2019-05-09
 ### Changed
-- Update Swift version to 5.0
+- Update Swift version to 5.0.
 
 ## [0.1.0] - 2019-05-09
 ### Added
 - `AppDelegate.swift`
 
+[0.2.0]: https://github.com/jasonnam/AppDelegate/compare/0.1.1...0.2.0
 [0.1.1]: https://github.com/jasonnam/AppDelegate/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/jasonnam/AppDelegate/releases/tag/0.1.0